댓글 삭제시 아이디 확인

+++ src/main/java/org/example/springboot/service/comments/CommentsService.java
@@ -1,6 +1,8 @@
 package org.example.springboot.service.comments;
 
 import lombok.RequiredArgsConstructor;
+import org.example.springboot.config.auth.LoginUser;
+import org.example.springboot.config.auth.dto.SessionUser;
 import org.example.springboot.domain.comments.Comments;
 import org.example.springboot.domain.comments.CommentsRepository;
 import org.example.springboot.web.dto.CommentsListResponseDto;
@@ -40,10 +42,16 @@ public class CommentsService {
 
 
     @Transactional
-    public void delete(Long id){
+    public void delete(Long id, @LoginUser SessionUser user){
         Comments comments = commentsRepository.findById(id)
                 .orElseThrow(()->new IllegalArgumentException("해당 댓글이 없습니다. id="+id));
-        commentsRepository.delete(comments);
+        if (user.getName().equals(comments.getAuthor())) {
+            commentsRepository.delete(comments);
+        }
+        else{
+            throw new IllegalAccessError("본인의 댓글만 삭제 가능합니다.");
+        }
+
     }
